backend/requirements.txt
fastapi
mangum
uvicorn[standard]
python-dotenv
pydantic
requests
yfinance
boto3